refactor (outputView) 출력포멧맞추기 - 테스트를 진행하면서 사용자 입력에서 parsing이 잘못된 부분들을 수정했습니다 - 영수증 출력에서 출력양식을 오른쪽 정렬하여 사용자 편의성을 높일 수 있도록 영수증 양식을 바꿨습니다

+++ src/main/java/store/service/ProductService.java
@@ -64,7 +64,7 @@ public class ProductService {
         }
         int giveaway = calculatePromoQuantity(quantity, product.getPromotion());
         // 프로모션 전량 적용 가능 시
-        if ((double) product.getQuantity() / promoThreshold == (int) product.getQuantity() / promoThreshold) {
+        if ((double) product.getQuantity() / promoThreshold == (double) (int) product.getQuantity() / promoThreshold) {
             return createReceipt(product, quantity, giveaway);
         }
